refactor:     - add base class for Layouter     - remove Spiral
