+++ TagsCloud/Layouter/CircularCloudLayouter.cs
@@ -1,70 +1,27 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Drawing;
-using TagsCloud.Extensions;
+using TagsCloud.Layouter.Base;
 
 namespace TagsCloud.Layouter
 {
-    public class CircularCloudLayouter : ILayouter
+    public class CircularCloudLayouter : LayouterBase
     {
-        private readonly List<Rectangle> rectangles;
-        private ISpiral spiral;
+        private readonly Point center;
 
-        public CircularCloudLayouter(ISpiral spiral)
+        public CircularCloudLayouter(Point center) : base(center)
         {
-            rectangles = new List<Rectangle>();
-            this.spiral = spiral;
+            this.center = center;
         }
 
-        public Rectangle PutNextRectangle(Size rectangleSize)
+        protected override IEnumerable<Point> GetPoints()
         {
-            var rectangle = new Rectangle();
-            rectangle.Size = rectangleSize;
-            foreach (var point in spiral.GetPoints())
+            for (var i = 0; ; i++)
             {
-                rectangle.Location = point;
-                if (!rectangle.IntersectsWith(rectangles))
-                    break;
+                var x = (int)(i * Math.Cos(i * 0.5)) + center.X;
+                var y = (int)(i * Math.Sin(i * -0.5)) + center.Y;
+                yield return new Point(x, y);
             }
-            var movedRectangle = MoveToCenter(rectangle);
-            rectangles.Add(movedRectangle);
-            return movedRectangle;
-        }
-
-        internal RectangleF[] ToArray()
-        {
-            throw new NotImplementedException();
-        }
-
-        private Rectangle MoveToCenter(Rectangle rectangle)
-        {
-            var targetVector = new TargetVector(spiral.Center, rectangle.Location);
-            foreach (var delta in targetVector.GetPartialDelta())
-            {
-                var newRectangle = rectangle.MoveOnTheDelta(delta);
-                if (newRectangle.IntersectsWith(rectangles))
-                    continue;
-                rectangle = newRectangle;
-            }
-            return rectangle;
-        }
-
-        public Size GetLayoutSize()
-        {
-            var size = new Size();
-            foreach (var rect in rectangles)
-            {
-                size.Width = Math.Max(rect.X - spiral.Center.X + rect.Width, size.Width);
-                size.Height = Math.Max(rect.Y - spiral.Center.Y + rect.Height, size.Height);
-            }
-            size.Height += spiral.Center.X * 2;
-            size.Width += spiral.Center.Y * 2;
-            return size;
-        }
-
-        public void SetCenter(Point center)
-        {
-            spiral.SetCenter(center);
         }
     }
 }
