+++ TagsCloud/Layouter/ILayouter.cs
@@ -4,8 +4,7 @@ namespace TagsCloud.Layouter
 {
     public interface ILayouter
     {
-        public Rectangle PutNextRectangle(Size rectangleSize);
-        public Size GetLayoutSize();
-        public void SetCenter(Point center);
+        Rectangle PutNextRectangle(Size rectangleSize);
+        Size GetLayoutSize();
     }
 }
